fix: repair two pre-existing build failures in tests and benches
teeny-cuda's elemwise_add test formatted `*mut f32` with `{:#x}`, which raw
pointers do not implement; cast to `usize` first.

teeny-kernels relied on a comment claiming `benches/conv2d_bn_silu.rs` was
"not registered", but with no `[[bench]]` section Cargo auto-discovered and
built it, failing on the deliberately-unlinked `nn::fused`. Turn off bench
auto-discovery so the file stays on disk as reference, as intended.

+++ drivers/teeny-cuda/tests/test_elemwise_add.rs
@@ -42,9 +42,9 @@ fn test_tensor_add() -> Result<()> {
     let out_buf = device.buffer::<f32>(N)?;
     println!(
         "[4/9] allocated device buffers: x={:#x} y={:#x} out={:#x}",
-        x_buf.as_device_ptr(),
-        y_buf.as_device_ptr(),
-        out_buf.as_device_ptr(),
+        x_buf.as_device_ptr() as usize,
+        y_buf.as_device_ptr() as usize,
+        out_buf.as_device_ptr() as usize,
     );
 
     x_buf.to_device(&x_host)?;
